Code init

+++ MicroRabbit.Banking.Domain/Commands/CreateTransferCommand.cs
@@ -10,7 +10,7 @@ namespace MicroRabbit.Banking.Domain.Commands
 {
     public class CreateTransferCommand : TransferCommand
     {
-        public CreateTransferCommand(Guid from, Guid to, double amount)
+        public CreateTransferCommand(Guid from, Guid to, decimal amount)
         {
             FromAccount = from;
             ToAccount = to;
