+++ MicroRabbit.Banking.Domain/Commands/TransferCommand.cs
@@ -11,6 +11,6 @@ namespace MicroRabbit.Banking.Domain.Commands
     {
         public Guid FromAccount { get; protected set; }
         public Guid ToAccount { get; protected set; }
-        public double Amount { get; protected set; }
+        public decimal Amount { get; protected set; }
     }
 }
